swdptap: Fix clock signal inconsistency during SWDIO turnaround
This commit rearranges the busy loop responsible for the low clock
timing generation. After this change, the low clock period is honored
as the debug probe becomes the receiver during a turnaround SWD cycle.

Prior to this commit, the low clock period had only been honored
when the debug probe becomes the transmitter during the turnaround
cycle. The low clock period had been omitted when the debug probe
becomes the receiver during a turnaround cycle.

Closes #